backend/app/model_stub.py
"""Model inference stub - replace with actual model implementation."""
import numpy as np
from typing import Dict, Tuple, Optional
from PIL import Image
import io


def inference(image_data: bytes) -> Tuple[Dict, np.ndarray, Optional[np.ndarray]]:
    """
    Run model inference on image data.
    
    Args:
        image_data: Raw image bytes
        
    Returns:
        Tuple of (predictions_dict, image_embedding, text_embedding)
        - predictions_dict: Contains room_type, condition_score, natural_light_score, feature_tags
        - image_embedding: numpy array of shape (768,) for image embedding
        - text_embedding: numpy array of shape (1536,) for text embedding (optional)
    """
    # Stub implementation - replace with actual model
    # In production, this would:
    # 1. Load image with PIL/cv2
    # 2. Preprocess for model
    # 3. Run ViT/CLIP model for embeddings
    # 4. Run multi-head classifier for room type, condition, etc.
    
    # Example: Load and validate image
    try:
        img = Image.open(io.BytesIO(image_data))
        img.verify()
    except Exception:
        # Return default values if image is invalid
        pass
    
    # Stub predictions
    predictions = {
        "room_type": {
            "label": "kitchen",
            "confidence": 0.93
        },
        "condition_score": 0.78,
        "natural_light_score": 0.61,
        "feature_tags": ["hardwood_floors", "island", "stainless_steel_appliances"]
    }
    
    # Stub embeddings - random vectors for now
    # Replace with actual model inference
    np.random.seed(42)  # For reproducibility in stub
    image_embedding = np.random.randn(768).astype(np.float32)
    image_embedding = image_embedding / np.linalg.norm(image_embedding)  # Normalize
    
    # Text embedding (optional - could be generated from caption)
    text_embedding = np.random.randn(1536).astype(np.float32)
    text_embedding = text_embedding / np.linalg.norm(text_embedding)
    
    return predictions, image_embedding, text_embedding


def generate_caption(image_data: bytes) -> str:
    """
    Generate a caption/alt-text for the image.
    
    Args:
        image_data: Raw image bytes
        
    Returns:
        Caption string
    """
    # Stub - replace with actual captioning model
    return "Modern kitchen with hardwood floors, center island, and stainless steel appliances."

